Backend/requirements.txt
# =========================
# CORE API
# =========================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# =========================
# DATABASE
# =========================
SQLAlchemy==2.0.30
pymysql==1.1.0

# =========================
# AUTH / SECURITY
# =========================
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
cryptography==42.0.5

# =========================
# VALIDATION
# =========================
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

# =========================
# FORMS / FILES
# =========================
python-multipart==0.0.9

# =========================
# EMAIL
# =========================
fastapi-mail==1.4.1

# =========================
# CORS / REQUESTS
# =========================
requests==2.31.0

# =========================
# TIMEZONE (Linux/Docker)
# =========================
tzdata>=2024.1
